+++ src/main/java/com/financialtargets/outflow/application/dto/EssentialOutflowUpdateDTO.java
@@ -5,20 +5,22 @@ import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.Positive;
 import org.hibernate.validator.constraints.Length;
 
+import java.math.BigDecimal;
+
 public record EssentialOutflowUpdateDTO(
         @JsonProperty("name")
         String name,
 
         @Positive
         @JsonProperty("value")
-        Float value,
+        BigDecimal value,
 
         @JsonProperty("dueDate")
         String dueDate,
 
         @Min(0)
         @JsonProperty("paidValue")
-        Float paidValue,
+        BigDecimal paidValue,
 
         @Length(max = 100)
         @JsonProperty("notes")
